AddressTranslator session->connection
